refactor(polyfills): add types
This change adds TypeScript type annotations to our polyfills. With
this change, the in-house polyfills will be shipped.
