+++ src/polyfills/toggle-class.ts
@@ -18,7 +18,7 @@ const missingNativeDOMTokenListToggleForce = (() => {
 if (missingNativeDOMTokenListToggleForce) {
   (() => {
     const origToggle = DOMTokenList.prototype.toggle;
-    DOMTokenList.prototype.toggle = function toggleDOMTokenList(name, add) {
+    DOMTokenList.prototype.toggle = function toggleDOMTokenList(name: string, add: boolean) {
       return arguments.length < 2 || this.contains(name) === !add ? origToggle.call(this, name) : add;
     };
   })();
